+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 * Fix issue where webhook verification would fail due to missing global public key configuration value
 * Fix issue where webhook testing did not work on instantiated gateway
+* Add support for VCR compelling evidence dispute representment
 
 ## 3.30.0
 * Add `oauthAccessRevocation` to `WebhookNotification`s